+++ scp/src/scpIfOperatorInterpreter.cpp
@@ -17,6 +17,12 @@ namespace scp
 ScAddrToValueUnorderedMap<std::function<std::unique_ptr<SCPOperator>(ScAgentContext &, ScAddr)>>
     ASCPIfOperatorInterpreter::supportedOperators = {};
 
+ASCPIfOperatorInterpreter::ASCPIfOperatorInterpreter()
+{
+  m_logger = utils::ScLogger(
+      utils::ScLogger::ScLogType::File, "logs/ASCPIfOperatorInterpreter.log", utils::ScLogLevel::Debug, true);
+}
+
 ScAddr ASCPIfOperatorInterpreter::GetActionClass() const
 {
   return Keynodes::action_interpret_if_operator;