+++ docs/changelog.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.2] -- 07.12.2025
+
+### Changed
+
+- Don't finish agent actions by ASCPHandlingEventThatInitiatesAgentSCPProgram
+- Set logging type FILE for all agents
+
+### Fixed
+
+- Duplicate result for agent actions
+
 ## [0.1.1] -- 06.12.2025
 
 ### Changed